Lg/yvlasiuk

See merge request gliesereum/gliesereum-platform!88

+++ lending-gallery/src/main/resources/db/migration/all/V2__add_tables_art_bond_dependensy.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS lending_gallery.art_bond (
    id uuid NOT NULL DEFAULT uuid_generate_v4(),
-   price bigint,
+   price integer,
    name character varying,
    description character varying,
    author character varying,
